+++ workalendar/europe/russia.py
@@ -1,11 +1,10 @@
-from ..core import OrthodoxCalendar
+from ..core import OrthodoxCalendar, MON
 from ..registry_tools import iso_register
 
 
 @iso_register('RU')
 class Russia(OrthodoxCalendar):
     'Russia'
-
     shift_new_years_day = True
 
     # Civil holidays
@@ -20,3 +19,15 @@ class Russia(OrthodoxCalendar):
         (6, 12, "National Day"),
         (11, 4, "Day of Unity"),
     )
+
+    def get_calendar_holidays(self, year):
+        holidays = super().get_calendar_holidays(year)
+        shifts = []
+        for day, label in holidays:
+            if day.weekday() in self.get_weekend_days():
+                shifts.append((
+                    self.get_first_weekday_after(day, MON),
+                    label + " shift"
+                ))
+        holidays.extend(shifts)
+        return holidays